requirements.txt
numpy>=1.24
pyyaml>=6.0
jsonschema>=4.19
spiceypy>=5.1.2
pytest>=7.4

